M2-PR-03: split Runtime Lifecycle tests into lifecycle/ subtree
Behaviour-preserving. 86 tests moved from tests/mod.rs to 10 leaf
files under tests/lifecycle/. After this PR, tests/mod.rs is 26 LOC —
just module declarations + re-exports; zero test bodies remain at
the directory root.

- lifecycle/fixtures.rs (67, 0): 4 backend-context helpers (pub(crate))
- lifecycle/restart_or_shutdown.rs (47, 3): rescued from
  "GNOME Extension State Parsing Tests" misnomer banner
- lifecycle/logind_decode.rs (388, 29): 9 rescued + 20 logind decode/
  path/error tests
- lifecycle/display_apply.rs (81, 5)
- lifecycle/runtime_target.rs (273, 10)
- lifecycle/persistent_dbus.rs (74, 2)
- lifecycle/sni_runtime.rs (311, 7)
- lifecycle/transition.rs (370, 9)
- lifecycle/provider.rs (98, 5)
- lifecycle/supervisor.rs (770, 16)

No production source files modified. cargo build clean; cargo test
--bin kanata-switcher 261/0. tests.rs split complete: 4864 LOC mono
file -> 22 leaf files + common/helpers.rs, all <= 770 LOC.
